refactor(common): move file-upload helpers out of utils into a service module
Tech-debt sweep finding 23.

common/utils.py imported common.tasks and common.models.FileUploadAudit at
module level while common/tasks.py imports common.models back, so
common/models/base.py and common/models/tags.py had to defer their
common.utils imports inside functions to avoid importing a half-initialised
common.models.

Move create_file_audit_and_scan, _validate_file_field and
safe_save_uploaded_file, the sole holders of both offending imports, into a
new common/service/upload_service.py. They belong in the service layer anyway:
they create audit rows, dispatch Celery scans and schedule thumbnails. Their
common.thumbnails imports become module-level there, since the reason for
deferring them is gone.

common/utils.py now imports neither common.models nor common.tasks and keeps
only import-light helpers (strip_exif, the race-protection wrappers,
update_db_instance, serve_image_or_placeholder). No compatibility re-export is
added, since it would recreate the exact module-level edge being deleted, so
every importer is updated, including the three string mock.patch targets that
would otherwise have broken silently.
