+++ src/common/tests/test_feature_malware_scan.py
@@ -4,16 +4,17 @@ from unittest import mock
 
 import pytest
 
-from common import tasks, utils
+from common import tasks
 from common.models import FileUploadAudit
+from common.service import upload_service
 
 
 @pytest.mark.django_db
 def test_scan_dispatched_when_flag_on(settings: t.Any, django_capture_on_commit_callbacks: t.Any) -> None:
     settings.FEATURE_MALWARE_SCAN = True
-    with mock.patch("common.utils.tasks.scan_for_malware.delay") as delay:
+    with mock.patch("common.service.upload_service.tasks.scan_for_malware.delay") as delay:
         with django_capture_on_commit_callbacks(execute=True):
-            utils.create_file_audit_and_scan(
+            upload_service.create_file_audit_and_scan(
                 app="events",
                 model="organization",
                 instance_pk=uuid.uuid4(),
@@ -29,8 +30,8 @@ def test_scan_dispatched_when_flag_on(settings: t.Any, django_capture_on_commit_
 @pytest.mark.django_db
 def test_scan_skipped_and_audit_clean_when_flag_off(settings: t.Any) -> None:
     settings.FEATURE_MALWARE_SCAN = False
-    with mock.patch("common.utils.tasks.scan_for_malware.delay") as delay:
-        utils.create_file_audit_and_scan(
+    with mock.patch("common.service.upload_service.tasks.scan_for_malware.delay") as delay:
+        upload_service.create_file_audit_and_scan(
             app="events",
             model="organization",
             instance_pk=uuid.uuid4(),
