+++ src/events/controllers/event_series_admin.py
@@ -10,9 +10,9 @@ from common.authentication import I18nJWTAuth
 from common.controllers import UserAwareController
 from common.models import Tag
 from common.schema import TagSchema, ValidationErrorResponse
+from common.service.upload_service import safe_save_uploaded_file
 from common.throttling import WriteThrottle
 from common.thumbnails.service import delete_image_with_derivatives
-from common.utils import safe_save_uploaded_file
 from events import models, schema
 from events.service import event_series_service, update_db_instance
