+++ src/events/controllers/organization_admin/core.py
@@ -7,9 +7,9 @@ from ninja_extra import api_controller, route
 from accounts.schema import VerifyEmailSchema
 from common.authentication import I18nJWTAuth
 from common.schema import EmailSchema, ErrorDetail, ValidationErrorResponse
+from common.service.upload_service import safe_save_uploaded_file
 from common.throttling import UserDefaultThrottle, WriteThrottle
 from common.thumbnails.service import delete_image_with_derivatives
-from common.utils import safe_save_uploaded_file
 from events import models, schema
 from events.controllers.permissions import IsOrganizationOwner, IsOrganizationStaff, OrganizationPermission
 from events.service import event_service, organization_service, stripe_service
